moved new samples to another directory
