+++ dataflow/gpu-examples/pytorch-minimal/.gcloudignore
