+++ dataflow/gpu-examples/pytorch-minimal/Dockerfile
