+++ dataflow/gpu-examples/tensorflow-landsat/.dockerignore
