+++ dataflow/gpu-examples/tensorflow-landsat/.gcloudignore
