+++ dataflow/gpu-examples/tensorflow-landsat/Dockerfile
