+++ dataflow/gpu-examples/tensorflow-minimal/.dockerignore
