+++ dataflow/gpu-examples/tensorflow-minimal/.gcloudignore
