+++ dataflow/gpu-examples/tensorflow-minimal/Dockerfile
