+++ dataflow/gpu-examples/tensorflow-minimal/build.yaml
