+++ dataflow/gpu-examples/tensorflow-minimal/noxfile_config.py
